+++ dsp_utils/Source/AudioHistoryBuffer.h
@@ -7,13 +7,13 @@ class AudioHistoryBuffer
 public:
     AudioHistoryBuffer(int historySize);
 
-    void pushSamples(const float* samples, int numSamples);
-    void getHistory(float* outSamples, int numSamples) const;
+    void pushSamples(const double* samples, int numSamples);
+    void getHistory(double* outSamples, int numSamples) const;
     int getSize() const { return size; }
-    void clear() { std::fill(buffer.begin(), buffer.end(), 0.0f); writePos = 0; }
+    void clear() { std::fill(buffer.begin(), buffer.end(), 0.0); writePos = 0; }
 
 private:
-    std::vector<float> buffer;
+    std::vector<double> buffer;
     int writePos = 0;
     int size = 0;
 };
